requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pyyaml>=6.0.1
httpx>=0.26.0
tenacity>=8.2.3
authlib>=1.3.0
python-multipart>=0.0.7
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0